It is alive!
If there are simulation results that you have not seen yet a new table shows up on the homepage and
displays the information on the simulations. It updates over time and will show if a simulation is still running
if it has completed or if it has failed. It contains the job id, when the job started and when it ended